backend/requirements.txt
django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg[binary]==3.2.9
